🚀 chore(openapi.json): update server URLs and title 🚀 chore(swagger.json): update host URL and title 🚀 chore(pom.xml): remove redundant name tag The server URLs in both openapi.json and swagger.json have been updated to use localhost instead of 127.0.0.1. The title in both files has been updated to use kebab-case instead of PascalCase. The name tag in pom.xml has been removed as it is redundant.
🚀 feat(OpenApiConfig.java): add servers to OpenAPI configuration
This commit adds two servers to the OpenAPI configuration, one for the Node application and one for the Auth application. The servers are defined using the `Server` class and added to the `OpenAPI` object using the `servers` method. This improves the flexibility of the application as it can now be deployed to different environments with different URLs.

🐛 fix(HomeController.java): change value of @Value annotation to "${info.application.artifactId}"
✨ feat(additional-spring-configuration-metadata.json): remove "application.name" property
✨ feat(application.properties): remove "info.application.name" property
✨ feat(test/resources/application.properties): remove "info.application.name" property

The changes in this commit remove the "application.name" property from the project's configuration metadata and properties files. The HomeController.java file has been updated to use the correct value for the @Value annotation, which is now "${info.application.artifactId}". This change ensures that the correct value is injected into the name variable.