[PAGOPA-681] Opex (#200)
* [PAGOPA-681] create dashboard

* [PAGOPA-681] opex

* [PAGOPA-681] provider

* [PAGOPA-681] terraform lock

* [PAGOPA-681] resource_type: api-management

* 🔀 chore(config.yaml): update data_source and resource_type
The data_source and resource_type fields in the prod environment configuration file have been updated. The data_source field now points to the application gateway instead of the API management service. The resource_type field has been commented out as it is no longer needed.

* 🚀 chore(openapi.json): update server URLs and title
🚀 chore(swagger.json): update host URL and title
🚀 chore(pom.xml): remove redundant name tag
The server URLs in both openapi.json and swagger.json have been updated to use localhost instead of 127.0.0.1. The title in both files has been updated to use kebab-case instead of PascalCase. The name tag in pom.xml has been removed as it is redundant.

🚀 feat(OpenApiConfig.java): add servers to OpenAPI configuration
This commit adds two servers to the OpenAPI configuration, one for the Node application and one for the Auth application. The servers are defined using the `Server` class and added to the `OpenAPI` object using the `servers` method. This improves the flexibility of the application as it can now be deployed to different environments with different URLs.

🐛 fix(HomeController.java): change value of @Value annotation to "${info.application.artifactId}"
✨ feat(additional-spring-configuration-metadata.json): remove "application.name" property
✨ feat(application.properties): remove "info.application.name" property
✨ feat(test/resources/application.properties): remove "info.application.name" property

The changes in this commit remove the "application.name" property from the project's configuration metadata and properties files. The HomeController.java file has been updated to use the correct value for the @Value annotation, which is now "${info.application.artifactId}". This change ensures that the correct value is injected into the name variable.

* 🔥 chore(context.json): remove context.json file
The context.json file has been removed from the repository as it is no longer needed.

* 🐛 fix(update_infra.yml): change openapi template file name from _openapi.json.tpl to _openapi.temp.json.tpl
✨ feat(update_infra.yml): remove second server from openapi template
The openapi template file name was changed from _openapi.json.tpl to _openapi.temp.json.tpl to avoid conflicts with the existing file. The second server was removed from the openapi template to ensure consistency with the infrastructure repository.

* multiple products

* 🚀 chore(create_dashboard.yaml): add concurrency configuration
This commit adds a concurrency configuration to the create_dashboard.yaml workflow file. The configuration specifies that jobs in the same workflow run should be grouped together and that in-progress jobs should not be cancelled when a new job is queued. This improves the efficiency of the workflow by allowing multiple jobs to run concurrently without interfering with each other.

* group concurrency

* 🚀 chore(create_dashboard.yaml): add Terraform Apply step and copy environments for auth product
This commit adds a Terraform Apply step to the create_dashboard.yaml workflow, which applies the Terraform configuration to the specified environment. Additionally, the environments for the auth product are now copied to the /azure-dashboard folder, allowing the Terraform Apply step to be executed for the auth product.

* 🐛 fix(config.yaml): add underscore to node config name
The node config name was missing an underscore at the end, which caused inconsistency with the naming conventions. The underscore has been added to the name to improve consistency.

🐛 fix(config.yaml): rename auth config name
The auth config name has been renamed to improve consistency with the naming conventions. The name has been changed from 'opex_pagopa-api-config' to 'opex_pagopa-api-config_auth'.

* fix

* fix

* fix