📦 chore(openapi): add .java-version file with Java version 11 A new file named .java-version has been added to the openapi directory. This file specifies that the Java version used for the project is version 11.
🔧 chore(CreditorInstitutionsController.java): organize imports and format code
The commit organizes the imports in the CreditorInstitutionsController.java file and formats the code for better readability. No functional changes were made.

🔧 chore(CreditorInstitutionsController.java): add support for filtering by enabled status
The CreditorInstitutionsController class now supports filtering by the enabled status of creditor institutions. A new request parameter named "enabled" has been added, allowing clients to filter the results based on whether the creditor institution is enabled or not.

🐛 fix(FilterPaView.java): add enabled field to FilterPaView class
🐛 fix(CreditorInstitutionsService.java): add enabled field to filter in CreditorInstitutionsService class
The `FilterPaView` class now includes an `enabled` field, allowing for filtering based on the enabled status of a PA view. The `CreditorInstitutionsService` class has been updated to include the `enabled` field in the filter when querying creditor institutions.

🐛 fix(PaStazionePaSpecification.java): fix null check for enabled parameter in PaStazionePaSpecification
✨ feat(PaStazionePaSpecification.java): add support for filtering by enabled parameter
The null check for the enabled parameter has been fixed to prevent potential NullPointerException. Additionally, support for filtering by the enabled parameter has been added, allowing for more specific queries based on the enabled status of the PaStazionePa entity.

🐛 fix(PaStazionePaSpecificationTest.java): update method calls with additional parameters
The method calls to `filterViewPaBrokerStation()` in the `PaStazionePaSpecificationTest` class have been updated to include additional parameters. This ensures that the method is called correctly with the updated signature and prevents any potential errors or inconsistencies.